T-9.3: convert to Arduino-as-ESP-IDF-component (IDF 5.5.5, arduino-esp32 3.3.11)
Firmware build of record moves from pioarduino/PlatformIO to ESP-IDF:
CMakeLists + main/ component, sdkconfig.defaults regenerated from the
Arduino core's own IDF-5.5.5 base + qio_opi overlay + T-9.3 deltas,
NB_TARGET env switch + tools/idf_build.sh replace the seven pio envs,
pack_web wired into the spiffs image target. PlatformIO keeps env:native.

mbedTLS ASYMMETRIC_CONTENT_LEN (16384/2048) + drop-peer-cert and the
64 KB data cache now resolve from source (they were inert at T-0.6/T-1.3);
libmbedtls.a is a from-source artifact. Runtime 53 KB re-measure awaits
hardware. Latent -Werror=all findings fixed at the source (card_test %s
format, %u/uint32_t casts, config.cpp restrict alias, render count-buffer
bound). lib/render: zero conversion changes; purity guard holds.

+++ src/envs/cache_test.cpp
@@ -20,7 +20,7 @@ using nb::data::kMaxGamesPerLeague;
 static void fill_generation(GameList* l, uint32_t gen) {
     l->count = 1 + static_cast<int>(gen % kMaxGamesPerLeague);
     char id[16];
-    std::snprintf(id, sizeof id, "%u", gen);
+    std::snprintf(id, sizeof id, "%lu", static_cast<unsigned long>(gen));
     for (int i = 0; i < l->count; i++) {
         Game& g = l->games[i];
         nb::data::copy_str(g.id, sizeof g.id, id);
@@ -38,7 +38,7 @@ static int check_generation(const GameList& l) {
     if (l.count < 1 || l.count > kMaxGamesPerLeague) return -1;
     const uint32_t gen = static_cast<uint32_t>(l.games[0].start_utc);
     char id[16];
-    std::snprintf(id, sizeof id, "%u", gen);
+    std::snprintf(id, sizeof id, "%lu", static_cast<unsigned long>(gen));
     if (l.count != 1 + static_cast<int>(gen % kMaxGamesPerLeague)) return static_cast<int>(gen);
     for (int i = 0; i < l.count; i++) {
         const Game& g = l.games[i];
